feat:relocate contexts and components folders
